merges HeFQUINEngine and  HeFQUINEngineImpl
